requirements.txt
selenium==4.10.0
pytz==2021.1
SQLAlchemy==2.0.16
PyMySQL==1.0.3
python-dotenv==1.0.0
sodapy==2.2.0
pandas==2.0.3
splinter==1.26.16
openpyxl==3.1.2
twilio==8.5.0
sendgrid==6.10.0